结构大调整，尽量降低与第三方ORM的耦合度。因为数字本身也是文字，所以删除整数表，添加字符串类型的数字表。

+++ src/main/java/cc/xfl12345/mybigdata/server/model/database/error/TableOperationException.java
@@ -27,7 +27,7 @@ public class TableOperationException extends Exception {
 
     protected String tableName;
 
-    protected int affectedRowsCount;
+    protected long affectedRowsCount;
 
     // public TableOperationException(String message, CURD operation, String tableName) {
     //     super(message);
@@ -35,7 +35,7 @@ public class TableOperationException extends Exception {
     //     this.tableName = tableName;
     // }
 
-    public TableOperationException(String message, int affectedRowsCount, CURD operation, String tableName) {
+    public TableOperationException(String message, long affectedRowsCount, CURD operation, String tableName) {
         super(message);
         this.affectedRowsCount = affectedRowsCount;
         this.operation = operation;
